Migrate http2 handling logic to Http2Connection trait.
